+++ bindings/python/candlewick/video_context.py
@@ -1,10 +1,22 @@
 from . import Visualizer
+from . import VideoRecorderSettings
 from contextlib import contextmanager
 
 
+_DEFAULT_VIDEO_SETTINGS = VideoRecorderSettings()
+
+
 @contextmanager
-def create_recorder_context(viz: Visualizer, filename: str, /, fps: int = 30):
-    viz.startRecording(filename, fps)
+def create_recorder_context(
+    viz: Visualizer,
+    filename: str,
+    /,
+    fps: int = _DEFAULT_VIDEO_SETTINGS.fps,
+    bitRate: int = _DEFAULT_VIDEO_SETTINGS.bitRate,
+):
+    viz.videoSettings().fps = fps
+    viz.videoSettings().bitRate = bitRate
+    viz.startRecording(filename)
     try:
         yield
     finally: